requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0
scikit-learn==1.3.0
numpy==1.24.3
python-dotenv==1.0.0
google-generativeai==0.3.0
pydantic